conversation creation and deletion implemented

+++ app/conversations/mutations/deleteConversation.ts
@@ -1,4 +1,4 @@
-import { resolver } from "blitz"
+import { AuthorizationError, resolver } from "blitz"
 import db from "db"
 import { z } from "zod"
 
@@ -9,8 +9,18 @@ const DeleteConversation = z.object({
 export default resolver.pipe(
   resolver.zod(DeleteConversation),
   resolver.authorize(),
-  async ({ id }) => {
+  async ({ id }, ctx) => {
     // TODO: in multi-tenant app, you must add validation to ensure correct tenant
+    const userToConversation = await db.userToConversation.findFirst({
+      where: { userId: ctx.session.userId, conversationId: id },
+    })
+    if (!userToConversation)
+      return new AuthorizationError("You lack permissions to delete this conversation.")
+
+    const userToConversationEntries = await db.userToConversation.deleteMany({
+      where: { conversationId: id },
+    })
+    const messages = await db.message.deleteMany({ where: { conversationId: id } })
     const conversation = await db.conversation.deleteMany({ where: { id } })
 
     return conversation
